requirement.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2<3.0,>=2.8.6
drf-spectacular>=0.15.1,<0.16
beautifulsoup4==4.10.0
requests==2.26.0